Add a behat feature about presigned requests
